challenge/requirements.txt
flask
pyjwt
playwright
